bichart: DataProcessor is refactored to easily put all calculation en separate thread

+++ bichart/src/java/com/biorecorder/bichart/scroll/ScrollModel.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class ScrollModel {
-    private double start = 0; // view start
-    private double end = 100; // view end (start-end are limits where move viewport)
+    private int start = 0; // view start
+    private int end = 100; // view end (start-end are limits where move viewport)
     private double viewportPosition = 0; // viewportPosition
-    private double viewportExtent = 10; // viewportWidth
+    private int viewportExtent = 10; // viewportWidth
 
     private List<ChangeListener> listeners = new ArrayList<ChangeListener>();
 
@@ -30,7 +30,7 @@ public class ScrollModel {
         return viewportPosition;
     }
 
-    public double getViewportExtent() {
+    public int getViewportExtent() {
         return viewportExtent;
     }
 
@@ -38,26 +38,26 @@ public class ScrollModel {
         setRangeProperties(newPosition, viewportExtent, start, end);
     }
 
-    public void setViewportExtent(double newExtent){
+    public void setViewportExtent(int newExtent){
         setRangeProperties(viewportPosition, newExtent, start, end);
     }
 
-    public void setStartEnd(double newStart, double newEnd) {
+    public void setStartEnd(int newStart, int newEnd) {
         setRangeProperties(viewportPosition, viewportExtent, newStart, newEnd);
     }
 
-    public double getStart() {
+    public int getStart() {
         return start;
     }
 
-    public double getEnd() {
+    public int getEnd() {
         return end;
     }
 
     /**
      * our model change limits (end) if extent is too big
      */
-    public void setRangeProperties(double newPosition, double newExtent, double newStart, double newEnd) {
+    public void setRangeProperties(double newPosition, int newExtent, int newStart, int newEnd) {
         double oldExtent = viewportExtent;
         double oldPosition = viewportPosition;
         end = newEnd;
@@ -84,7 +84,7 @@ public class ScrollModel {
         return extent;
     }*/
 
-    private double normalizeValue(double value, double extent) {
+    private double normalizeValue(double value, int extent) {
         if (value < start) {
             return start;
         }
